txn: rename txn_complete_async to txn_on_journal_write
The function is called only by the journal when write is finished.

Besides, it may not complete the transaction. In case of
synchronous replication it is not enough for completion. It means,
it can't have 'complete' in its name.

Also the function is never used out of txn.c, so it is removed
from txn.h and is now static.

The patch is a preparation for not spaming "too long WAL write" on
synchronous transactions, because it is simply misleading.

Part of #5139

(cherry picked from commit a512cb8ab1e388377a5dfdbd9c13f07e3ff245e1)