vclock: move to src/lib
Vclock is used in raft, which is going to be moved to src/lib.
That means vclock also should be moved there.

It is easy, because vclock does not depend on anything in box/.

Needed for #5303

(cherry picked from commit b8d3c762b284c36c7ba85f1dc4d4607c594e4ffb)