raft: move synchro queue clear to the worker fiber
The synchro queue was cleared from the raft on_update trigger
installed by box. It was fine as long as the trigger is called
from the worker fiber, because it won't block the state machine,
while the synchro queue clearance yields.

But the trigger is going to be called from the raft state machine
directly soon. Because it will need to call
box_update_ro_summary() right after raft state is updated, without
a yield to switch to the worker fiber. This will be done in scope
of getting rid of box in the raft library.

It means, the trigger can't call box_clear_synchro_queue(). But it
can schedule its execution for later, since the worker fiber now
belongs to box. The patch does it.

Part of #5303

(cherry picked from commit 63d312eea2d41d7395c84bb43ce1473cf6e5edfe)