raft: move RO summary update to box-Raft
RO summary update is supposed to make the instance read-only, when
it becomes a follower, and read-write when becomes a leader.

But it makes raft depend on box, and prevents raft move to a
separate library.

The patch moves the RO update to box-raft.

This became possible after some preparatory work was done to make
raft update triggers non-yielding, and invoked right after state
change (without a yield between the change and the triggers).

Part of #5303