test: update test-run (--snapshot)
* Added --snapshot and --disable-schema-upgrade arguments (#240).
* Fixed reporting of an error for conflicting arguments (#241).

The `--snapshot path/to/snapshot` argument copies a given snapshot to a
snapshot directory before start a tarantool instance. This allows to
verify various functionality in the case, when tarantool is upgraded
from a snapshot that is left by an older tarantool version (as opposite
to test it on a freshly bootstrapped instance). There are limitations:
when a test spawns a replica set, the option does not work correctly.
The reason is that the same instance UUIDs (and IDs) cannot be used by
different instances in a replica set. Maybe there are other pitfalls.

The `--disable-schema-upgrade` argument instructs tarantool to skip
execution of the schema upgrade script (using ERRINJ_AUTO_UPGRADE). This
way we can verify that, when an instance works on an old schema version,
a functionality is workable or at least gives correct error message.

This commit only brings the new options into test-run. It does NOT add
any new testing targets / rules.

Part of #4801

(cherry picked from commit af0e365e2cdae8881c5bcd35b11d22eba01477d8)