test: update test-run (handle no result file)
This changeset fixes a problem that unlikely will hit anyone, but in
theory it may be triggered by an incorrect behaviour of tarantool.

In brief, if tarantool does not react to SIGTERM after executing all
tests and a test-run's worker stucks at waiting for termination of the
tarantool process, the test-run's listener would fail at attempt to
access a temporary result file that does not exists. See more details in
[1].

[1]: https://github.com/tarantool/test-run/issues/245