test: update test-run (SIGKILL for hang tarantool)
Limit waiting for a tarantool process termination by 5 seconds. When
this timeout exceeded, print a warning to the terminal and send SIGKILL
to the process.

We need to handle the situation with a stuck tarantool process on the
testing system side to overcome a problem of this kind that appears on
Mac OS (see #5573).

This changeset handles one particular case: stopping of a tarantool
instance that either started for execution of a 'core = tarantool' test
suite or started from a test using the `test_run:cmd('start server
foo')` command. It does not handle stopping of tarantool that is started
for execution of a 'core = app' test or started from a test directly
using io.popen() or built-in 'popen' module.

Related to #5573
Part of https://github.com/tarantool/test-run/issues/157
The changeset: https://github.com/tarantool/test-run/pull/186